requirements.txt
asgiref==3.5.2
autopep8==2.0.1
dj-database-url==1.2.0
Django==4.1.4
django-cors-headers==3.13.0
django-js-asset==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.3.0
pycodestyle==2.10.0
PyJWT==2.6.0
python-decouple==3.6
pytz==2022.6
sqlparse==0.4.3
whitenoise==6.2.0
